ben-add: logic to check if epic exists already
very janky way of writing it, will likely update later but this works for now. May want to do a useState for it and set it by checking if the current deck has an epic. If it does not have an epic, allow a new epic to come, otherwise not allow it