ben-update: collection endpoints to allow wildcards
necessary to ensure when a new user who is not logged in has a way to track that it is a new, unregistered deckbuilder. new tag will be used to know to save the deck in local storage and then save later when they login/make a new account. Forces users to go to a new collection if they are not signed in, so logic needs to be in place on load for collection/deckbuilder to check if they are logged in or not