ben-add: logic to show card count updates on +/- butt press
also fixed bug when searching first and then clicking. i needed to find the correct index to update because a search changes the size of the array, so the array index is not the value to search based on card num. Instead, i needed to search for the index based on the object's cardNum value, hence why findIndex() func was used