ben-add: context for decklist value, WIP
currently added decklist, is empty on default and needs to hit API to insert starting data eventually. Currently gets called in card.tsx component and will be needed to be added in all places where we add to decklist. Will aslso need to worry about removing next and a func is already made, not finished yet